♻️ 7A3K50 integrate: decompose merge helpers
