+++ .agentplane/tasks/202605282126-7A3K50/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 7A3K50 task: PR integrate merge strategy decomposition [202605282126-7A3K50]
