+++ .agentplane/tasks/202605282126-7A3K50/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605282126-7A3K50/pr-integrate-merge-strategy-decomposition",
+  "created_at": "2026-05-28T21:26:47.182Z",
+  "diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "last_verified_at": "2026-05-28T21:30:26.507Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605282126-7A3K50",
+  "updated_at": "2026-05-28T21:26:47.182Z",
+  "verify": {
+    "status": "pass"
+  }
+}
