+++ .agentplane/tasks/202605282126-7A3K50/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T21:26:47.182Z
+
+## Task
+
+- Task: `202605282126-7A3K50`
+- Title: PR integrate merge strategy decomposition
+- Status: DOING
+- Branch: `task/202605282126-7A3K50/pr-integrate-merge-strategy-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282126-7A3K50/README.md`
+
+## Verification
+
+- State: ok
+- Note: PR integrate merge strategy decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T21:26:47.182Z
+- Branch: task/202605282126-7A3K50/pr-integrate-merge-strategy-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
